Replace ArrayStack entirely
